requirements.txt
flask
transformers
torch
Pillow
requests
matplotlib
